Significantly reduce the number of re-renders on the console page when stats are flowing
